+++ runtime/src/lib.rs
@@ -30,6 +30,7 @@ pub mod prioritization_fee_cache;
 mod read_optimized_dashmap;
 pub mod rent_collector;
 mod reward_info;
+pub use reward_info::RewardInfo;
 pub mod runtime_config;
 pub mod serde_snapshot;
 pub mod slot_params;